BlogBoT/requirements.txt
Flask
google.generativeai
flask_cors
newsapi-python
google-api-python-client
pymongo
gunicorn
